feat: Support `Uint8List` and `File` types as inputs
Extends the functionality of `generatePDFFromDocuments`, `mergeMultiplePDFs`, and `createPDFFromMultipleImages` to accept a list of `dynamic` inputs, which can now be `String` paths, `Uint8List` byte arrays, or `File` objects.

A new private helper method, `getPathsFromInputs`, is introduced to handle the conversion of these dynamic inputs into a unified list of file paths. For `Uint8List`, temporary files are created to be processed.

Additionally, this commit updates the macOS deployment target from 10.14 to 10.15 in the example project's `Podfile` and Xcode project settings.